requirements.txt
flask
flask-cors
streamlit
google-generativeai
